miniraft

Module debug

Module for pretty printing state transitions, log updates, etc. No actual Raft-specific logic.

Structs

Logger

Wrapper struct that contains methods for logging specific program flows in Raft

Enums

AnnotationType

Internal debug message to dump contents of entries and state

Level

Level of logging

Functions

assertion

Helper function to log a test check

colour_bool

Helper function to pretty print a boolean

colour_server

Helper function to pretty print a ServerId with a unique colour

colour_term

Helper function to pretty print a Term

debug_log

Pretty print a set of Annotations over a vector of LogEntries

init_logger

Initialize the logger (default uses microseconds)

log

Helper function to pretty print a message at the corresponding log Level

Type Aliases

Annotation

A tuple representing an annotation message and its AnnotationType