miniraft::log

Struct Log

pub struct Log<T, S> {
    pub entries: Vec<LogEntry<T>>,
    pub committed_len: LogIndex,
    pub applied_len: LogIndex,
    pub app: Box<dyn App<T, S>>,
    pub parent_id: ServerId,
}

A collection of LogEntries

Fields
entries: Vec<LogEntry<T>>

Log entries

committed_len: LogIndex

How much of the log has been considered committed. A log entry is considered ‘safely replicated’ or committed once it is replicated on a majority of servers. Only meaningful on servers which are leaders. Initialized to 0 as no entries are committed. Increases monotonically.

applied_len: LogIndex

How much of the log has been applied to the state machine. Initialized to 0, increases monotonically.

app: Box<dyn App<T, S>>

State machine

parent_id: ServerId

ServerId of our parent for pretty printing documentation

Implementations

impl<T, S> Log<T, S>
where T: Debug,

pub fn new(parent_id: ServerId, app: Box<dyn App<T, S>>) -> Self

Instantiate a new empty event log

pub fn last_term(&self) -> Term

Fetch the most recent term we have recorded in the log

pub fn last_idx(&self) -> LogIndex

Get index of the last element

pub fn append_entries( &mut self, prefix_idx: LogIndex, leader_commit_len: LogIndex, entries: Vec<LogEntry<T>>, )

Append additional entries to the log. prefix_idx is what index caller expects entries to be inserted at, leader_commit_len is the index of last log that leader has commited.

pub fn deliver_msg(&mut self)

Deliver a single message from the message log to the application