miniraft::log

Trait App

pub trait App<T, S> {
    // Required methods
    fn transition_fn(&mut self, entry: &LogEntry<T>);
    fn get_state(&self) -> S;
}

Describes a state machine that is updated bassed off of a feed of LogEntry

Required Methods

fn transition_fn(&mut self, entry: &LogEntry<T>)

Function that mutates the application state depending on the newest log entry. Raft guarantees that if the transition function is called on a LogEntry, it is considered applied (meaning it won’t be re-run or removed).

fn get_state(&self) -> S

Return the current state of the application