miniraft

Module server

Module containing majority of the logic for handling RPCs, managing state transitions, and the API

Structs

CandidateState

Candidate specific volatile state

FollowerState

Follower specific volatile state

LeaderState

Leader specific volatile state

NodeReplicationState

State of a single Node as tracked by a leader

RaftConfig

Configuration options for a Raft server

RaftServer

A Raft server that replicates Logs of type T

Enums

RaftLeadershipState

Possible states a Raft Node can be in

Type Aliases

ServerId

Type alias for the ID of a single Raft server

Term

Type alias for Raft leadership term