Struct miniraft::server::RaftServer

pub struct RaftServer<T, S> {
    pub id: ServerId,
    pub current_term: Term,
    pub log: Log<T, S>,
    /* private fields */
}

A Raft server that replicates Logs of type T

Fields
id: ServerId

ID of this node

current_term: Term

Current term of this node

log: Log<T, S>

List of log entries for this node. This is the data that is being replicated

Implementations

impl<T, S> RaftServer<T, S>where T: Clone + Debug,

pub fn new( id: ServerId, peers: BTreeSet<ServerId>, config: RaftConfig, seed: Option<u64>, app: Box<dyn App<T, S>> ) -> Self

Create a new Raft node with a given ID. Caller is responsible for ensuring it is unique. Initialize with all peers in the cluster along with an App that runs over the event log to arrive at a state.

pub fn tick(&mut self) -> Vec<SendableMessage<T>>

Tick state and perform necessary state transitions/RPC calls

pub fn quorum_size(&self) -> usize

Calculate quorum of current set of peers. quorum = ceil((peers.length + 1)/2)

pub fn receive_rpc(&mut self, rpc: &RPC<T>) -> Vec<SendableMessage<T>>

Demultiplex incoming RPC to its correct receiver function

pub fn client_request(&mut self, msg: T) -> Result<()>

Public interface for clients to request adding log entries to the cluster. Will fail if the node it is called on a non-Leader node

pub fn promote_to_leader( &mut self, followers: BTreeMap<ServerId, NodeReplicationState> ) -> Vec<SendableMessage<T>>

Manually promote node to leader. Do not call during normal operation.

pub fn is_leader(&self) -> bool

Logging helpers /// Whether current node is a Leader

pub fn is_candidate(&self) -> bool

Whether current node is a Candidate

pub fn is_follower(&self) -> bool

Whether current node is a Follower