Struct miniraft::debug::Logger

pub struct Logger {}

Wrapper struct that contains methods for logging specific program flows in Raft

Implementations

impl Logger

pub fn append_entries_recv<T: Debug, S>(
log_ref: &Log<T, S>,
prefix_idx: LogIndex,
leader_commit_len: LogIndex,
their_entries: &Vec<LogEntry<T>>
)

called when a node receives a request to append entries

pub fn log_potential_conflict<T: Debug, S>(
log_ref: &Log<T, S>,
their_entries: &Vec<LogEntry<T>>,
prefix_idx: LogIndex,
rollback_to: LogIndex
)

called on potential log conflict when appending entries

pub fn log_term_conflict<T: Debug, S>(log_ref: &Log<T, S>)

detected a term conflict, log details about truncation

pub fn log_append<T: Debug, S>(log_ref: &Log<T, S>, start: LogIndex)

details about actually appending to the log

pub fn log_apply<T: Debug, S>(log_ref: &Log<T, S>, leader_commit_len: LogIndex)

details about applying a number of log entries to the state machine

pub fn log_deliver_recv<T: Debug, S>(log_ref: &Log<T, S>)

called when delivering a single log entry to the application

pub fn log_deliver_apply<T: Debug, S>(log_ref: &Log<T, S>)

called when log entries are done being applied to state machine (application)

pub fn server_init<T: Debug + Clone, S>(raft_ref: &RaftServer<T, S>)

initializing a server

pub fn state_update<T: Debug + Clone, S>(raft_ref: &RaftServer<T, S>)

log a leadership state transition

pub fn won_election<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
num_votes: usize,
follower_ids: &Vec<ServerId>
)

log election states upon winning

pub fn send_heartbeat<T: Debug + Clone, S>(raft_ref: &RaftServer<T, S>)

leader sending heartbeat to followers

pub fn election_timer_expired<T: Debug + Clone, S>(raft_ref: &RaftServer<T, S>)

candidate/follower election timeout reached

pub fn outgoing_rpcs<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
msgs: Vec<SendableMessage<T>>
) -> Vec<SendableMessage<T>>

log single outgoing rpc request (including type and target)

pub fn check_matching_term<T>(
id: &ServerId,
req: &AppendRequest<T>,
current_term: Term
)

rpc request pre-req: ensure term matches before continuing

pub fn bumping_term<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
new_term: Term
)

log when a term change/update has occurred

pub fn receive_rpc<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
rpc: &RPC<T>
)

log incoming rpc request (including type and received from)

pub fn client_request<T: Debug + Clone, S>(raft_ref: &RaftServer<T, S>)

log client API calls

pub fn replicate_entries<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
entries: &Vec<LogEntry<T>>,
target: &ServerId,
prefix_len: LogIndex
)

log when leader prepares to replicate log entries to followers

pub fn rpc_vote_request<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
req: &VoteRequest
)

follower receiving a request from a candidate to vote for them

pub fn rpc_vote_result<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
log_ok: bool,
up_to_date: bool,
havent_voted: bool
)

explain follower decision making for whether to vote for candidate

pub fn rpc_vote_resp<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
res: &VoteResponse
)

candidate receiving a vote result from a follower

pub fn vote_count(id: &ServerId, res: &VoteResponse, up_to_date: bool)

log decision making process for candidate deciding whether result from follower is valid

pub fn total_vote_count(id: &ServerId, total: usize, quorum: usize)

log total votes for candidate

pub fn added_follower<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
votee: &ServerId
)

log adding a follower under a leader

pub fn rpc_append_request<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
req: &AppendRequest<T>
)

log when follower receives a request to append log entries from leader

pub fn append_conflict_check<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
req: &AppendRequest<T>
)

checking for potential log conflict before appending

pub fn append_entries<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
prefix_ok: bool,
last_log_entry_matches_terms: bool,
prefix_len: usize
)

log follower appending entries from leader

pub fn append_response<T: Debug + Clone, S>(
raft_ref: &RaftServer<T, S>,
res: &AppendResponse
)

log leader receiving response from follower re: append_entries

pub fn process_append_response(
id: &ServerId,
res: &AppendResponse,
follower_state: &NodeReplicationState
)

log decision making process for leader when updating its replication state for a follower

pub fn commit_entry(
id: &ServerId,
commit_len: LogIndex,
acks: usize,
quorum_size: usize
)

log decision making process on a leader about whether to commit entries