Module miniraft::rpc

Module containing definitions for all of the RPCs that Raft nodes use to communicate with each other.

Structs

AppendRequest

Request from leader to append entries to follower’s log

AppendResponse

Response to an AppendRequest

VoteRequest

Request by a candidate to become a Raft leader

VoteResponse

Response to a VoteRequest

Enums

RPC

A Raft RPC request

Target

Whether to send a message to everyone or just a single node

Type Definitions

SendableMessage

A message can be either targeted at a single server or to everyone